third_party/WebKit/LayoutTests/fast/html/imports/sub-imports-onload-expected.txt

This ensures that load and error events are fired on sub-imports.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS theEvent.type is 'error'
PASS theEvent.target.ownerDocument is root.import
PASS theEvent.type is 'load'
PASS theEvent.target.ownerDocument is root.import
PASS successfullyParsed is true

TEST COMPLETE
